feat: altera o código para javascript
